+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - 2018-07-27
+### Added
+- `load_from_json`- a command that provides an ability to define config settings
+  by loading them from a json file (in `load_from_yaml` manner);
+
+### Changed
+- Support for Ruby 2.2 has ended;
+
 ## [0.4.0] - 2018-06-24
+### Added
 - Introduce Plugin Ecosystem (`Qonfig::Plugins`):
   - load plugin: `Qonfig.plugin('plugin_name')` or `Qonfig.plugin(:plugin_name)`;
   - get registered plugins: `Qonfig.plugins #=> array of strings`